@@ -11,11 +11,12 @@
   "author": "John Clark <john@treehub.com>",
   "license": "MIT",
   "devDependencies": {
-    "eslint": "3.15.0",
+    "eslint": "3.19.0",
     "eslint-config-google": "0.7.1",
-    "lerna": "2.0.0-beta.37",
+    "lerna": "2.0.0-rc.2",
     "rollup": "0.41.6",
     "rollup-plugin-node-resolve": "3.0.0",
-    "rollup-plugin-string": "2.0.2"
+    "rollup-plugin-string": "2.0.2",
+    "rollup-watch": "3.2.2"
   }
 }
