@@ -1,6 +1,6 @@
-const core = require('th-core');
+import attr from '@thp/mixins/attr';
 
-class Component extends core(HTMLElement) {
+class Component extends attr(HTMLElement) {
   constructor() {
     super({
       attributes: Component.observedAttributes,
@@ -14,6 +14,11 @@ class Component extends core(HTMLElement) {
     ];
   }
 
+  connectedCallback() {
+    super.connectedCallback();
+    this._asideChanged(this.aside, null);
+  }
+
   _asideChanged(newAside, oldAside) {
     // TODO don't recreate when package is the same
     const path = decodeURI(newAside).split('/').filter((s)=>s !== '');
@@ -22,18 +27,15 @@ class Component extends core(HTMLElement) {
       return this.innerHTML = '';
     }
     const pkg = window.packages[path.shift().toLowerCase()];
-    if (pkg == undefined) {
+    if (pkg === undefined) {
       return this.innerText = 'Package not installed';
     }
 
-    if (pkg.aside == undefined) {
+    if (pkg.aside === undefined) {
       return this.innerText = 'Package has no aside';
     }
     this.innerHTML = `<${pkg.aside} path="/${path.join('/')}"></${pkg.aside}>`;
   }
 }
 
-const template = require('./template.html');
-document.head.insertAdjacentHTML('beforeend', template);
-
-window.customElements.define('th-aside', Component);
+export default Component;
